+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 - updated pykit@0.8.0
 - listed public exports
 - renamed throw_err_evt() -> throw()
-- removed ServerBusCfg.is_invoked_action_unhandled_errs_logged in favor of ServerBusCfg.are_errs_catchlogged
+- added ServerBusCfg.trace_errs_on_pub which allows to catch+reraise thrown to the bus errs
+- removed ServerBusCfg.is_invoked_action_unhandled_errs_logged in favor of ServerBusCfg.trace_errs_on_pub
 - ServerBusCfg: renamed register -> register_fn
 - setup ContextVar for subfn subscriber call
 - added support for custom subfn context functions